backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
pydantic-settings>=2.4,<3.0
PyJWT>=2.9,<3.0
psycopg[binary]>=3.2,<4.0
pytest>=8.3,<9.0
httpx>=0.27,<1.0
